+++ Changelog.md
@@ -10,6 +10,7 @@ _yyyy.mm.dd_
 * Remove support for cycles in trained TPG graphs. _This change alters the deterministic outputs of trainings._ As explained in [Issue #106](https://github.com/gegelati/gegelati/issues/106), the genetic mutation of the graph, as described by Stephen Kelly in his PhD manuscript will never introduce any cycle in the trained TPG. As a consequence, parts of the code have been simplified:
   * Mutations: It is no longer mandatory to have an action referenced in each team of the TPG.
   * TPG Execution (in gegelati and in generated code): When executing outgoing edges of a team, providing a mechanism to exclude already visited edges is not needed.
+* Add the attribute "bestScoreLastGen" to the learning agent that records the best score achieved in each generation. This score can be obtained using the "getBestScoreLastGen" method.
 
 ### Bug fix
 * Fix build issue with MSVC19. `#include <ctime>` needed in timestamp.cpp.